+++ src/main/java/wooteco/subway/admin/controller/LineController.java
@@ -10,11 +10,11 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.admin.domain.Line;
-import wooteco.subway.admin.dto.LineRequest;
-import wooteco.subway.admin.dto.LineResponse;
-import wooteco.subway.admin.dto.LineStationAddRequest;
-import wooteco.subway.admin.dto.LineStationResponse;
-import wooteco.subway.admin.dto.LineUpdateRequest;
+import wooteco.subway.admin.dto.request.LineRequest;
+import wooteco.subway.admin.dto.request.LineStationAddRequest;
+import wooteco.subway.admin.dto.request.LineUpdateRequest;
+import wooteco.subway.admin.dto.response.LineResponse;
+import wooteco.subway.admin.dto.response.LineStationResponse;
 import wooteco.subway.admin.service.LineService;
 
 import java.net.URI;
