+++ src/main/java/wooteco/subway/admin/dto/request/LineStationAddRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.request;
 
 public class LineStationAddRequest {
     private String lineName;
