+++ src/main/java/wooteco/subway/admin/service/LineService.java
@@ -4,9 +4,9 @@ import org.springframework.stereotype.Service;
 import wooteco.subway.admin.domain.Line;
 import wooteco.subway.admin.domain.LineStation;
 import wooteco.subway.admin.domain.Station;
-import wooteco.subway.admin.dto.LineResponse;
-import wooteco.subway.admin.dto.LineStationAddRequest;
-import wooteco.subway.admin.dto.LineStationResponse;
+import wooteco.subway.admin.dto.request.LineStationAddRequest;
+import wooteco.subway.admin.dto.response.LineResponse;
+import wooteco.subway.admin.dto.response.LineStationResponse;
 import wooteco.subway.admin.repository.LineRepository;
 import wooteco.subway.admin.repository.StationRepository;
