+++ src/test/java/wooteco/subway/admin/acceptance/LineAcceptanceTest.java
@@ -9,7 +9,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import wooteco.subway.admin.dto.LineResponse;
+import wooteco.subway.admin.dto.response.LineResponse;
 
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
